renamed files and splitted carousel files
- updated carousel service button style
