+++ data/events/pop-up-data.ts
@@ -2,4 +2,5 @@ export const event = {
   title: 'DataStax Roadshow',
   img: 'sctc-popup.jpg',
   link: '/Roadshow',
+  enabled: false,
 };
